[CA-6] Refactored repository into  a multi module maven project with modules for api and data. Created README.md file with a sample checlist.
