+++ twitter-clone-server/README.md
@@ -1,0 +1,12 @@
+# TwitterClone
+A Twitter Clone App for our MPP course at Maharishi University of Management.
+
+### Project Checklist
+#### Server Side
+
+    [x] Spring Boot with SpringMVC
+    [x] REST API
+
+#### Client Side
+    [x] ReactJS
+
