+++ twitter-clone-server/twitter-clone-api/src/main/java/com/mpp/twitterclone/TwitterCloneServerApplication.java
@@ -1,4 +1,4 @@
-package com.mpp.twittercloneserver;
+package com.mpp.twitterclone;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
