+++ twitter-clone-server/twitter-clone-api/src/main/resources/application.properties
