+++ twitter-clone-server/twitter-clone-api/src/test/java/com/mpp/twitterclone/TwitterCloneServerApplicationTests.java
@@ -1,4 +1,4 @@
-package com.mpp.twittercloneserver;
+package com.mpp.twitterclone;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
